Important Update: Changes in Root Detection
Dear Wi-Fi KeyShare (ROOT) User Community,

We are pleased to announce a significant update in Wi-Fi KeyShare (ROOT) related to the detection of root access on Android devices. Starting from the latest version, we have adopted a more transparent and accessible approach to implement this feature.

Changes in Root Detection:

Instead of using a third-party library, we have integrated a shell to perform real-time root access checking. This decision is rooted in our commitment to transparency and user privacy.

What Does This Mean for You as a User?

What sets this update apart is that you can now view the specific command used to verify root access directly from our public repository. We have chosen to share this information as part of our effort to provide clarity on how our applications work and how this detection is carried out.

This means that if you have concerns or questions about root access detection in Wi-Fi KeyShare (ROOT), you can refer to the code in our public repository to see the exact command being executed. We believe that this transparency is essential in building trust with our users.

If you have any questions or comments about this update, please feel free to contact our support team. We appreciate your ongoing support and contribution to making Wi-Fi KeyShare (ROOT) a better application for everyone.

Thank you for being a part of our community.